% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_QA_mask.R
\name{mini_QA_mask}
\alias{mini_QA_mask}
\title{QA mask for minitree}
\usage{
mini_QA_mask(
  DTM_fp,
  outdir,
  mini_segment_sf_list,
  slope_thres = 45,
  sos_thres = 84.5,
  save_mask = F
)
}
\arguments{
\item{DTM_fp}{Filepath to the DTM. A DTM of 1m ground resolution was used in Thapa et al. (2026)}

\item{out_dir}{Output directory.}
}
\value{
A List with two objects. (1) sf object showing the treetops (POINTS) and (2) sf object showing the segmented crowns (POLYGONS)
}
\description{
A function that takes the digital terrain model (DTM) as input and returns recommended QA masks
}
