% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mini_segment.R
\name{mini_segment}
\alias{mini_segment}
\title{Segmenting small trees}
\usage{
mini_segment(shm_r)
}
\arguments{
\item{shm_r}{A spatRaster of canopy heights created by the lidarSHM pipeline, created by increasing max_shrub_ht to maximum height of trees}
}
\value{
A List with two objects. (1) sf object showing the treetops (POINTS) and (2) sf object showing the segmented crowns (POLYGONS)
}
\description{
A function to create segment trees using LiDAR point clouds as an input. Specifically designed to be sensitive to small regenerating trees.
}
